requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchainhub==0.1.1
python-dotenv==0.21.0
numexpr
yfinance>=0.2.40
duckduckgo-search>=6.0.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu
pypdf
mysql-connector-python